Fix up QHeaderView failing on the default nodejs MS compiler

+++ src/cpp/lib/QtWidgets/QHeaderView/qheaderview_wrap.cpp
@@ -28,7 +28,7 @@ QHeaderViewWrap::QHeaderViewWrap(const Napi::CallbackInfo& info)
   Napi::Env env = info.Env();
   Napi::HandleScope scope(env);
 
-  int len = info.Length();
+  size_t len = info.Length();
 
   this->disableDeletion = false;
   if (len == 1) {
